backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy==2.0.36
python-multipart>=0.0.9
